core/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
fonttools==4.53.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
paypalrestsdk==1.13.3
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pyfiglet==1.0.2
pygame==2.6.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyTelegramBotAPI==4.20.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.0
stripe==10.8.0
telebot==0.0.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
